requirements.txt
Django==3.0.5
django-cors-headers==3.3.0
django-cors-middleware==1.5.0
djangorestframework==3.11.0
numpy==1.17.2
pandas==0.25.1
requests==2.22.0
sklearn==0.0
utils==0.9.0
webencodings==0.5.1
websocket-client==0.56.0